Serverless/Backend/pyproject.toml
[tool.poetry]
name = "lora-dashboard-backend"
version = "0.1.0"
description = "FastAPI backend for LoRA Dashboard - Serverless Training & Generation Suite"
authors = ["Mateusz <mateusz@example.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
redis = "^5.0.1"
boto3 = "^1.34.0"
pyyaml = "^6.0.1"
asyncio-subprocess = "^0.1.0"
python-multipart = "^0.0.6"
httpx = "^0.25.2"
psutil = "^5.9.6"
aiofiles = "^23.2.1"
runpod = "^1.7.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
ruff = "^0.1.6"
mypy = "^1.7.1"
httpx = "^0.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
target-version = "py312"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
disallow_any_generics = true
disallow_untyped_defs = true
follow_imports = "silent"
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_unimported = true 